refactor(python): Move core objects to top level
