+++ py-polars/polars/dataframe/__init__.py
@@ -1,0 +1,6 @@
+from polars.dataframe.frame import DataFrame, wrap_df
+
+__all__ = [
+    "DataFrame",
+    "wrap_df",
+]
