+++ py-polars/polars/dataframe/frame.py
@@ -25,6 +25,8 @@ from typing import (
 )
 
 from polars import internals as pli
+from polars.dataframe._html import NotebookFormatter
+from polars.dataframe.groupby import DynamicGroupBy, GroupBy, RollingGroupBy
 from polars.datatypes import (
     FLOAT_DTYPES,
     INTEGER_DTYPES,
@@ -65,8 +67,6 @@ from polars.internals.construction import (
     sequence_to_pydf,
     series_to_pydf,
 )
-from polars.internals.dataframe._html import NotebookFormatter
-from polars.internals.dataframe.groupby import DynamicGroupBy, GroupBy, RollingGroupBy
 from polars.internals.io_excel import (
     _unpack_multi_column_dict,
     _xl_apply_conditional_formats,
