--- py-polars/polars/internals/dataframe/__init__.py
@@ -1,6 +1,0 @@
-from polars.internals.dataframe.frame import DataFrame, wrap_df
-
-__all__ = [
-    "DataFrame",
-    "wrap_df",
-]
