--- py-polars/polars/internals/series/__init__.py
@@ -1,6 +1,0 @@
-from polars.internals.series.series import Series, wrap_s
-
-__all__ = [
-    "Series",
-    "wrap_s",
-]
