+++ py-polars/polars/lazyframe/__init__.py
@@ -1,0 +1,6 @@
+from polars.lazyframe.frame import LazyFrame, wrap_ldf
+
+__all__ = [
+    "LazyFrame",
+    "wrap_ldf",
+]
