+++ py-polars/polars/series/__init__.py
@@ -1,0 +1,6 @@
+from polars.series.series import Series, wrap_s
+
+__all__ = [
+    "Series",
+    "wrap_s",
+]
