+++ py-polars/polars/series/datetime.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import polars.internals as pli
-from polars.internals.series.utils import expr_dispatch
+from polars.series.utils import expr_dispatch
 from polars.utils.convert import _to_python_datetime
 from polars.utils.decorators import deprecated_alias, redirect
