+++ py-polars/polars/series/string.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import polars.internals as pli
-from polars.internals.series.utils import expr_dispatch
+from polars.series.utils import expr_dispatch
 
 if TYPE_CHECKING:
     from polars.datatypes import PolarsDataType, PolarsTemporalType
