+++ py-polars/tests/unit/test_constructors.py
@@ -264,9 +264,7 @@ def test_init_ndarray(monkeypatch: Any) -> None:
         _ = pl.DataFrame(np.array([[1, 2], [3, 4]]), schema=["a"])
 
     # NumPy not available
-    monkeypatch.setattr(
-        pl.internals.dataframe.frame, "_check_for_numpy", lambda x: False
-    )
+    monkeypatch.setattr(pl.dataframe.frame, "_check_for_numpy", lambda x: False)
     with pytest.raises(ValueError):
         pl.DataFrame(np.array([1, 2, 3]), schema=["a"])
 
@@ -488,9 +486,7 @@ def test_init_pandas(monkeypatch: Any) -> None:
     assert df.rows() == [(datetime(2022, 10, 31, 10, 30, 45, 123456),)]
 
     # pandas is not available
-    monkeypatch.setattr(
-        pl.internals.dataframe.frame, "_check_for_pandas", lambda x: False
-    )
+    monkeypatch.setattr(pl.dataframe.frame, "_check_for_pandas", lambda x: False)
     with pytest.raises(ValueError):
         pl.DataFrame(pandas_df)
